protocol: Move Js5InboundChannelHandler to protocol package
Matches it's function within the pipeline
